refactor: add parachain detection
Simply attempts to build as a normal Rust project otherwise.
