+++ crates/pop-common/Cargo.toml
@@ -9,6 +9,7 @@ repository.workspace = true
 
 [dependencies]
 anyhow.workspace = true
+cargo_toml.workspace = true
 git2.workspace = true
 git2_credentials.workspace = true
 regex.workspace = true
