+++ crates/pop-contracts/Cargo.toml
@@ -12,7 +12,6 @@ repository.workspace = true
 
 [dependencies]
 anyhow.workspace = true
-cargo_toml.workspace = true
 duct.workspace = true
 flate2.workspace = true
 reqwest.workspace = true
