+++ crates/pop-contracts/src/build.rs
@@ -1,23 +1,18 @@
 // SPDX-License-Identifier: GPL-3.0
 
-use crate::utils::helpers::get_manifest_path;
+use crate::{errors::Error, utils::helpers::get_manifest_path};
 use contract_build::{execute, BuildMode, ExecuteArgs};
-use std::path::{Path, PathBuf};
-use thiserror::Error;
-
-#[derive(Error, Debug)]
-pub enum Error {
-	#[error("Failed to get manifest path: {0}")]
-	ManifestPath(String),
-	#[error("Manifest error: {0}")]
-	ManifestError(#[from] cargo_toml::Error),
-}
+use std::path::Path;
 
 /// Build the smart contract located at the specified `path` in `build_release` mode.
-pub fn build_smart_contract(path: Option<&Path>, build_release: bool) -> anyhow::Result<String> {
+///
+/// # Arguments
+/// * `path` - The optional path to the smart contract manifest, defaulting to the current directory if not specified.
+/// * `release` - Whether the smart contract should be built without any debugging functionality.
+pub fn build_smart_contract(path: Option<&Path>, release: bool) -> anyhow::Result<String> {
 	let manifest_path = get_manifest_path(path)?;
 
-	let build_mode = match build_release {
+	let build_mode = match release {
 		true => BuildMode::Release,
 		false => BuildMode::Debug,
 	};
@@ -32,19 +27,9 @@ pub fn build_smart_contract(path: Option<&Path>, build_release: bool) -> anyhow:
 }
 
 /// Determines whether the manifest at the supplied path is a supported smart contract project.
+///
+/// # Arguments
+/// * `path` - The optional path to the manifest, defaulting to the current directory if not specified.
 pub fn is_supported(path: Option<&Path>) -> Result<bool, Error> {
-	// Resolve manifest path
-	let path = match path {
-		Some(path) => match path.ends_with("Cargo.toml") {
-			true => path.to_path_buf(),
-			false => path.join("Cargo.toml"),
-		},
-		None => PathBuf::from("./Cargo.toml"),
-	};
-	if !path.exists() {
-		return Err(Error::ManifestPath(path.display().to_string()));
-	}
-	let manifest = cargo_toml::Manifest::from_path(path)?;
-	// Simply check for the `ink` dependency
-	Ok(manifest.dependencies.contains_key("ink"))
+	Ok(pop_common::manifest::from_path(path)?.dependencies.contains_key("ink"))
 }
