+++ crates/pop-parachains/src/errors.rs
@@ -5,57 +5,42 @@ use zombienet_sdk::OrchestratorError;
 
 #[derive(Error, Debug)]
 pub enum Error {
-	#[error("Failed to access the current directory")]
-	CurrentDirAccess,
-
-	#[error("Failed to locate the workspace")]
-	WorkspaceLocate,
-
-	#[error("Failed to create pallet directory")]
-	PalletDirCreation,
-
-	#[error("IO error: {0}")]
-	IO(#[from] std::io::Error),
-
-	#[error("HTTP error: {0}")]
-	HttpError(#[from] reqwest::Error),
-
-	#[error("Missing binary: {0}")]
-	MissingBinary(String),
-
-	#[error("Configuration error: {0}")]
-	Config(String),
-
-	#[error("Unsupported command: {0}")]
-	UnsupportedCommand(String),
-
-	#[error("ParseError error: {0}")]
-	ParseError(#[from] url::ParseError),
-
-	#[error("Orchestrator error: {0}")]
-	OrchestratorError(#[from] OrchestratorError),
-
-	#[error("Toml error: {0}")]
-	TomlError(#[from] toml_edit::de::Error),
-
-	#[error("Anyhow error: {0}")]
-	AnyhowError(#[from] anyhow::Error),
-
 	#[error("User aborted due to existing target folder.")]
 	Aborted,
-
-	#[error("Failed to execute rustfmt")]
-	RustfmtError(std::io::Error),
-
-	#[error("Template error: {0}")]
-	TemplateError(#[from] pop_common::templates::Error),
-
-	#[error("Failed to parse the endowment value")]
-	EndowmentError,
-
-	#[error("Unsupported platform: {arch} {os}")]
-	UnsupportedPlatform { arch: &'static str, os: &'static str },
-
+	#[error("Anyhow error: {0}")]
+	AnyhowError(#[from] anyhow::Error),
 	#[error("Archive error: {0}")]
 	ArchiveError(String),
+	#[error("Configuration error: {0}")]
+	Config(String),
+	#[error("Failed to access the current directory")]
+	CurrentDirAccess,
+	#[error("Failed to parse the endowment value")]
+	EndowmentError,
+	#[error("HTTP error: {0}")]
+	HttpError(#[from] reqwest::Error),
+	#[error("IO error: {0}")]
+	IO(#[from] std::io::Error),
+	#[error("Manifest error: {0}")]
+	ManifestError(#[from] pop_common::manifest::Error),
+	#[error("Missing binary: {0}")]
+	MissingBinary(String),
+	#[error("Orchestrator error: {0}")]
+	OrchestratorError(#[from] OrchestratorError),
+	#[error("Failed to create pallet directory")]
+	PalletDirCreation,
+	#[error("ParseError error: {0}")]
+	ParseError(#[from] url::ParseError),
+	#[error("Failed to execute rustfmt")]
+	RustfmtError(std::io::Error),
+	#[error("Template error: {0}")]
+	TemplateError(#[from] pop_common::templates::Error),
+	#[error("Toml error: {0}")]
+	TomlError(#[from] toml_edit::de::Error),
+	#[error("Unsupported command: {0}")]
+	UnsupportedCommand(String),
+	#[error("Failed to locate the workspace")]
+	WorkspaceLocate,
+	#[error("Unsupported platform: {arch} {os}")]
+	UnsupportedPlatform { arch: &'static str, os: &'static str },
 }
